backend/requirements.txt
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.9.0
firebase-admin==2.13.0
google-api-core==1.8.2
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
idna==2.8
msgpack==0.6.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
